Server/requirements.txt
spotipy==2.23.0
python-dotenv
Flask